ReadMe.txt and FeedMe.IQ
I am a huge influence in the idea that inspiration comes when you are active and not while your eating chips on the couch. I also believe that reliance on money to sustain your cyber liberties is futile, It may help keep your computer working, but in hard times it will leave you defenseless.